Un-export fields
